[releng] Add support for GEF 3.15
Change from Import-Package to Require-Bundle as GEF 3.15 move some
code we depend on in a new package. We would need to import the new
package to build without error against 3.15, but that would cause a
failure on 3.14, so the only way to be compatible with both is to
switch to Require-Bundle.

Change-Id: I525bad1def1f3437c87187f03f264d19760d4f2d
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>