[releng] Declare an explicit dependency to Xalan
Apache Batik 1.16 no longer depends on Xalan, but unfortunately the
Orbit version of batik-bridge and batik-dom still declare
this (unneeded) dependency.

Until this is fixed in Orbit and/or GMF Runtime, add an explicit
dependency in the corresponding Sirius plug-in and feature so that
projects which use Sirius make sure they include Xalan to avoid
built-time or runtime issues.

Change-Id: I2e43b02d72dd2a9d02da5f97e82d97895c3bf385
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>