[581429] Always add Paste and Copy actions in tabbar
* Before this commit, the Paste actions are visible in the tabbar only
when no diagram element is selected, and the Copy action is visible in
the tabbar only when at least one diagram element is selected.
* The method TabbarPasteFormatMenuManager.update() has been modified
because the status of actions (inside the drop-down menu) were not
correctly refreshed. Even if the dropdown was enabled, when you expanded
the drop-down menu, the actions were always disabled after the first
Copy.
* The tooltip is now changed depending on whether diagram elements are
selected or not.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=581429
Change-Id: If9df297444fc13237d5471b7c4b9c4370d4f8c47