[581336] Fix o.e.s.tests.swtbot.tree.TreeUIPermissionAuthorityTests
In commit [1], there is a regression detected by the tests from
org.eclipse.sirius.tests.swtbot.tree.TreeUIPermissionAuthorityTests.
This regression concerns the permission authority and the fact that the
tree item retrieves its previous collapse/expand state if the permission
authority forbids the collapse/expand. This regression is not really
visible by end-user. Indeed, these tests go through a path not really
accessible by end-user.

This commit correctly handles the case covered by the automatic tests,
but not representative of the real life.

In conclusion for end-user:
* With or without this commit, the graphical collapse/expand is reverted
if it is done with the left or right key.
* With or without this commit, the graphical collapse/expand is not
reverted if it is done with a mouse click. This problem already exists
before bug 581336 and bug 580934 [2].

[1] https://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=93e62a724a8c6c636973913137b7565a31af2acb
[2] https://bugs.eclipse.org/bugs/show_bug.cgi?id=580934

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=581336
Change-Id: I1773d731a1382323679c4b4261467160e3ef3059