[580524] Fix issue because of diff/merge
Consider the use case using diff/merge between P1 and P2.
If Project1 has an opened Session, when diff/merge between P1 and P2,
the change done in P2 will be notified in
UpdateImageDependenciesPreCommitListener of the P1 Session. It is ugly.
Diff/merge should put P2 resources in a dedicated ResourceSet.
But, for now this commit add a workaround which is to ignore
notification which notifier is not in P1 resources.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=580524
Change-Id: I2d182b8ea6de7d126a018c6fa02c61a2d9086e51
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>