[582091] Do not apply center layout on Region content at creation time
System property to enable the behavior:
-Dorg.eclipse.sirius.diagram.ui.internal.region.content.canonical.layout.standard=true

When the Region is created in the same operation, it will be layouted in
its RegionContainer. If the regions container is created it will be
centered or placed where the user clicked.
The goal of the "centered" layout is to regroup "externally" created
views (by the precommit canonical refresh) when they have no location
hint in the SiriusDataLayoutManager. This is done to ease the retrieval
of such elements created as secondary effect of a user action. The main
created view would take the clicked location as hint.

This seems to be unnecessary when there are several level of created
views. The top level created elements will be centered, but centering
their center can appear to require more layout work by the user than a
first default arrange. This also allows to have a consistent layout for
those elements compared to a diagram creation.

Bug: https://eclip.se/582091
Cherry-picked-from: https://eclip.se/582089
Change-Id: I18f3f3fb08f74fee614dda43155b12731e0b67cc
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>Sir