[582091] Use model constraint to layout Regions
System property to enable the behavior:
-Dorg.eclipse.sirius.diagram.ui.internal.region.autosize.figure.optimized=true

We used to store the computed constraint during layout, but this make the
figures unable to adapt their size to the content in auto-size.
Behavior is wrong for the user: if he close and reopen the diagram (or
do some undo redo, or add new content) : the layout changes.

Bug: https://eclip.se/582091
Cherry-picked-from: https://eclip.se/582089
Change-Id: I9f28c92a705b549277a6e8534c1fb72ba3f38728
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>