[582091] Enable individual auto-size action and status for Regions
System property to enable the behavior:
-Dorg.eclipse.sirius.diagram.ui.internal.region.autosize.action.optimized=true

When auto-size is called on RegionContainer : it put all regions in
auto-size.

When auto-size is called on a single regions (or selected regions), it
enables auto-size for the selection.

Without the system property: no individual status.action, everything
is delegated to the RC.

Bug: https://eclip.se/582091
Cherry-picked-from: https://eclip.se/582089
Change-Id: I4283cdd588647b81b79aa44588f91ad05adbd7f8
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>