[581811] Correctly reconnect Note Attachments after changes
This commit also corrects when a node is deleted after a D&D, with the
"remove note when attached element is removed" preference enabled, the
note is now properly removed.

Note: This preference now behaves the same for notes and texts when
deleting elements.

Note: `test_NoteAttachments_deletion` has been deleted because it
expects to behave in the opposite way to the ticket.

Bug: 581811
Change-Id: I5f094714cae8650c67d51220b2a21a5886be0437
Signed-off-by: Séraphin Costa <seraphin.costa@obeosoft.com>