[82] Fix disconnected note attachment problem with edge and note/text
Fix the indirect deletion of note attachment, when
deleting note/text.

Fix the indirect hiding of hiding note attachment when hiding edges.

This commit also corrects when a note/text is deleted or edge is hide,
with the "remove/hide note when attached element is removed" preference
enabled, the note attached indirectly is now properly removed or hidden.
Preference now works the same way for text in corrected cases.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/82
Signed-off-by: Séraphin Costa <seraphin.costa@obeosoft.com>