[145] Arrange linked border nodes when updating their container
Fix a layout issue when nodes or container with linked border nodes are
updated (i.e. with a semantic DnD): border nodes are now properly laid
out.
The constant SiriusLayoutDataManager.KEEP_FIXED has been added  to tag
elements to be laid out and indicate that their location should not be
modified.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/145
Change-Id: Ia4c011b6d9eb46956c7176cfdb0c5874087eaa8f
Signed-off-by: Glenn Plouhinec <glenn.plouhinec@obeo.fr>