[test] Fix EdgeLabelsMoveFromEdgeMoveTest regression since GMF 1.16.1
Some tests from
org.eclipse.sirius.tests.swtbot.EdgeLabelsMoveFromEdgeMoveTest fail
since Sirius upgraded to GMF 1.16.1(commit e4d058cf [1]). More
precisely,
they fail since the GMF commit [2] "[33] Increase the precision in edge
segments computation" to fix the Github issue gmf-runtime 33 [3].

As the computation has been slightly changed on the GMF side, the tests
must be adapted. The test data has also been slightly changed. The node
"A" and the node "B" has been moved and then move again to the same
location. We can observe that the bendpoints and the label locations are
slightly different (because of the GMF changes). I think a migration
participant should have been added when moving to GMF 1.16.1.

This commit also changes the EdgeLabelQuery. A new case has been
detected where vector and referenceVector are the same, but the angle is
not 0 but a double near 0. This change reinforces what was done for
Bugzilla 465328 [4].

[1] https://github.com/eclipse-sirius/sirius-desktop/commit/e4d058cf5a513672d4dfb7d9e73ea352832c15ba
[2] https://github.com/eclipse/gmf-runtime/commit/b4316c675b70d17f47129c0f2bf57000d49defe9
[3] https://github.com/eclipse/gmf-runtime/issues/33
[4] https://bugs.eclipse.org/bugs/show_bug.cgi?id=465328