[144][test] Fix a memory leak regarding tests failures
During an execution of a suite, if some fail during the failureTearDown
method execution, all the data used by these tests leak until the end of
the suite execution.
This can be problematic in case of long suite with several failures and
lot of data. It increases the risk of OutOfMemory error.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/144