[176] Adapt SiriusDiagramActionBarContributor to avoid memory leak
Create a specific ActionBarContributor for each specific implementation
of SiriusDiagramEditor to avoid a duplication of ActionRegistry: one for
"SiriusDiagramEditor.class" and one for real implementation
("DDiagramEditorImpl.class" or "EcoreEntitiesReadOnlyEditor.class").
Indeed, only one of the ActionRegistry is correctly cleaned at the
closing of the editor.

To avoid a performance problem, the timeout
"sun.awt.datatransfer.timeout" has been set to 100 ms instead of 10000
ms for the automatic tests (see comment on issue [1] for more details).

[1] https://github.com/eclipse-sirius/sirius-desktop/issues/176#issuecomment-2020256957

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/176