[205] Add tool to copy-paste style for any diagram elements
This commit also removes the existing Apply Appearance Properties
action, which does the same thing but in a different way.

This commit also renames the existing paste style, layout and format.

This commit also fix some test that fail due to this new feature.

Note that some changes on AbstractSiriusSwtBotGefTestCase.java are due
to the maximum file size set by CheckStyle.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/205
Signed-off-by: Séraphin Costa <seraphin.costa@obeosoft.com>