[65] Update DNode size computation and allows auto-size
For SquareDescription, it is now possible to produce nodes in
auto-size mode: their size will self-adapt to the label content (when
the label has the Node label position).

To enable this mode on a node mapping:
- it must have a Square Style Description with label position set to
"node" (or the resulting node having a Square style and the label
position set by the user to "node")
- the width and height must be set to -1
- the size computation must not be empty: its result will give the
default size dimension to use as starting point for the figure before
adapting the label.

Minimum width and height are both equals to 0 for all other node styles.

Previous behavior to evaluate/use/handle sizeExpression only when
width/height have 0,0 is modified to evaluate/use/handle the expression
when the value is (0,0) or (-1,-1). (0,-1) and (-1, 0) will for the
moment produce the same result than (-1, -1).

Default value for sizeExpression is still
NodeStyleDescriptionImpl.SIZE_COMPUTATION_EXPRESSION_EDEFAULT = "3";

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/65
Signed-off-by: Glenn Plouhinec <glenn.plouhinec@obeo.fr>
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>