[65] Adapt the size of nodes to their label
A new LayoutManager has been created to adapt the size of nodes to their
label. This LayoutManager is used if the label is displayed on the node
and if the size computation expression is set to "-1".
DNodeQuery.isAutoSizeStyle() method is used to avoid replacing the GMF
size with a default size, and instead keep the -1/-1 size as for
containers.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/65
Change-Id: I7265925af4bebdcaf1ea28e2bff5909ae2eca31c
Signed-off-by: Glenn Plouhinec <glenn.plouhinec@obeo.fr>