[203] Adapt code to ELK changes
For issue, https://github.com/eclipse/elk/issues/651, some
UnsupportedGraphException are now thrown if the configuration is invalid
(done in PR https://github.com/eclipse/elk/pull/885). It is the case in
Sirius as the root graph has not the CROSSING_MINIMIZATION_STRATEGY set
to INTERACTIVE, as it is the case for the container. Since ELK 0.9.0, it
seems that the CONSIDER_MODEL_ORDER_STRATEGY is more adapted to keep
order of children (instead of INTERACTIVE used for that before).

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/203