[152] Add new color-changing menus
A new popup has been defined to choose the color to use for the selected
element. This ColorPalettePopup is invoked from the Appearance tab in
the Properties View. Four color categories has been defined in this
popup:
- Last Used colors category displays the ten last colors used by the
user. These colors are persisted in preferences.
- Custom colors category displays the ten first colors defined by the
user. These colors are persisted in the aird file.
- Suggested colors category displays the ten "preferred colors" defined
by the user which are a subset of the colors defined in the VSM. These
colors are persisted in the aird file.
- Basic colors category displays ten useful colors.
This popup is invoked when clicking on the "Font Color", "Fill Color"
and "Line Color" buttons. Each of these properties has specific
properties or annotation to distinct colors used between these
properties.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/152
Signed-off-by: Glenn Plouhinec <glenn.plouhinec@obeo.fr>
Signed-off-by: lfasani <laurent.fasani@obeo.fr>