[18] Avoid to store too much data in SiriusLayoutDataManager
With the last commit [1] concerning this issue, too much views are
considered as to be centered. For example, when a representation is
created without opening it (case of representations created at the
Viewpoint selection). This leads to a memory leak. Indeed,
SiriusLayoutDataManagerImpl.addCreatedViewForInitPositionLayout(Diagram,
LinkedHashSet<View>) was called, but
SiriusLayoutDataManagerImpl.removeLayoutViews(Diagram) was not call.

It seems that we can avoid to store data in method
"DiagramDialectServices.createRepresentation". Indeed, the flag
NotYetOpenedDiagramAdapter has just been added before and in this case
the stored data is ignored in
SiriusCanonicalLayoutHandler.launchArrangeCommand(DiagramEditPart).

A test has also been modified to check the detected leak (to avoid
further similar regression).

[1] https://github.com/eclipse-sirius/sirius-desktop/commit/b8a5c6a9d0888e4f26a99210f70bd247a9981c51

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/18