[205] Complete anti NPEs fix
A first preventive fix was made in commit
2ccb31c44715da9c89af6ed080b1e671d845f260 by changing listeners of
PinElementsAction and PasteStylePureGraphicalAction to update their
action state with a displaySyncExec call instead of a displayAsyncExec
in order to prevent the action from being disposed of due to an
asynchronous call.

In the commit complete the preventive with a better handling of
potentially null active page.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/205
Change-Id: I04232b9626c7fb2a0cbf992fad8339bd2fdd5da0
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>