[333] Store edge routing in parent
During the conversion from DDiagram to ELK Graph, the routing styles of
original Sirius DEdge is now stored on the common parent of edges in ELK
Graph.
This property can be overridden by defining the routing style directly
in the VSM.

The tests have also been adapted. Indeed, before this commit the routing
applied was always Rectilinear. Now, it is not the case. The existing
tests have been "duplicated", to have the previous behavior (by forcing
the "Rectilinear routing" in the VSM) and the new behavior for all
existing tests by considering the initial used routing style.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/333