[331] Complete an existing migration participant
The previous version of the migration participant does not handle all
the cases. The case of the image style of nodes contained in a container
(and not directly in diagram) was missed.
The version number of the migration participant has just been bumped as
there is no incidence to call this migration on an aird file already
migrated.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/331