[343] Refresh workspace after export actions
If the exported file is in the workspace, this last is refreshed (with
the action launched in a WorkspaceModifyOperation to avoid collision
problem).
The workspace's refresh is not launched during the creation of the files
during the arrange all, to avoid potential side effects. In this case,
the refresh will still have to be done manually.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/343