[362] Fix the bug in the refresh of the layer tooltip
When the VSM is changed, the tooltip of all layers is updated. But the
tooltip list was never cleared between two menuShow. So the first N
tooltip text (with N, the number of layers) were always the same and the
updated tooltip were in an unreachable position (at the end) in the
list.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/362
Signed-off-by: Séraphin Costa <seraphin.costa@obeosoft.com>