[test] Fix RefreshWithCustomizedStyleTests
Tests from RefreshWithCustomizedStyleTests fail on some CI servers. It
seems that the field is not "validated" and the model/session is not
modified.

This commit adds several conditions to make tests more reliable:
- a "traverse" call for text fields to force the validation for this
kind of field,
- the Properties view is maximized before setting fields. At least the
modification of a combo box can be problematic on Linux if the field is
not "visible",
- a condition, OperationDoneCondition, has been added. This allows to
have a better feedback in case of failure,
- several "SWTBotUtils.waitAllUiEvents();" have been added.