[205] Correctly clean clipboard from SetStyleToWorkspaceImageTests
Some tests from
org.eclipse.sirius.tests.swtbot.SetStyleToWorkspaceImageTests copy an
image to the clipboard but do not clean it after the test. This could
have side effects on the following tests.

This was the case in some tests of TabBarTest that failed in the suite
org.eclipse.sirius.tests.swtbot.suite.AllUnreliableTestSuite because
SetStyleToWorkspaceImageTests is executed before TabBarTest.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/205