[394] Add variables "diagram" and "view" during direct edit execution
For a direct edit tool, the variables "diagram" and "view" were
available during the evaluation of `inputLabelExpression` but they were
no longer available during the execution of the tool.
This commit solves this problem by setting these variables just before
setting of edit mask variables.

By the way, all the edit mask variables are never cleaned. This is not
the goal of this issue.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/394