[396] Adapt NodeCreationEditPolicy to compute the target location
- Constructors of CreationUtil have been improved to directly have
LayoutData as parameter instead of location and size. Indeed, the
location is relative to the parent edit part and for border nodes, the
parent edit part is not systematically the "host" of the called policy.
- In NodeCreationEditPolicy, in case of border nodes, the "host edit
part" is not systematically considered (to compute the RootLayoutData
for example). The correct edit part, a parent of the host, is computed
according to the mapping defined in the used tool.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/396