[426] Fix border nodes tests regression
Analysis: It seems that before, the code with comment "manage location
of bordered node with closest side", in GMFHelper.getAbsoluteLocation,
is called only in specific circumstances. Since recent changes, it was
called systematically, but that shouldn't be the case.

This commit also reduces the number of getAbsoluBounds methods to ease
the impact analysis.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/426