[DRAFT] Commit to analyze refreshOnOpening pref value
On a specific CI server, some sequence diagram tests fail. One of them
fails with a wrong value of refreshOnOpening preference:
org.eclipse.sirius.tests.swtbot.sequence.MessageExtensionTest.testSessionSyncOnDiagramWithEdgemappingImportOpening
junit.framework.AssertionFailedError: The refresh on opening should be
enabled for this test.
	at junit.framework.Assert.fail(Assert.java:57)
	at junit.framework.Assert.assertTrue(Assert.java:22)
	at junit.framework.TestCase.assertTrue(TestCase.java:192)
	at org.eclipse.sirius.tests.swtbot.sequence.MessageExtensionTest.testSessionSyncOnDiagramWithEdgemappingImportOpening(MessageExtensionTest.java:114)
	at ...

This commit is here to try to understand why this preference as a wrong
value. A test probably sets this preference and lets it in a wrong
state.