[test] Fix some tests failing since "preferences management change"
Some tests fail since the commit dc87bccb "[test] improves the
restoration of preferences in the test
" [1]. Its goals was to allow better restoration of the initial
preference values when using "changeXXXPreference" methods of
SiriusTestCase or AbstractSiriusSwtBotGefTestCase superclasses. The side
effect is that some tests were OK before because they relied on "wrong"
default values ​​for some preferences.

This commit fixes some of these tests. The main effect was that
SaveSessionJob is called and fails as it is not really expected. With
the preference PREF_SAVE_WHEN_NO_EDITOR set to false, the SaveSessionJob
is no longer called. But sometimes the cause-effect link, for the same
cause/same preference, is less clear.

Some other tests are maybe also impacted by the commit [1]. This commit
will be maybe followed by other ones. It is difficult to identify all as
the result depends on the context.

[1] https://github.com/eclipse-sirius/sirius-desktop/commit/dc87bccb24c6637c0b7a842a4baa99a4a89253b5