[test] Fix LayerTests failing since "prefs management change"
This test fails since the commit dc87bccb "[test] improves the
restoration of preferences in the test" [1] on some CI.

After a long analysis, it seems that the SaveSessionJob caused by the
"genericSetUp(SEMANTIC_MODEL_PATH, MODELER_PATH);" [2] can handle to
conflict with the changes made in "initViewpoint(VIEWPOINT_NAME);" [3] .
For example, sometimes the notification of the creation of the
"DSemanticDiagramSpec" is not "thrown". I think this problem may be
cross-cutting across multiple tests. It was probably OK before because
the preference "PREF_SAVE_WHEN_NO_EDITOR" was false and so no
SaveSessionJob was launched.

[1] https://github.com/eclipse-sirius/sirius-desktop/commit/dc87bccb24c6637c0b7a842a4baa99a4a89253b5
[2] https://github.com/eclipse-sirius/sirius-desktop/blob/8054e3ca60408e539c0590c8fbdbed713fe5b30e//plugins/org.eclipse.sirius.tests.junit/src/org/eclipse/sirius/tests/unit/diagram/modeler/uml/LayerTests.java#L60
[3] https://github.com/eclipse-sirius/sirius-desktop/blob/8054e3ca60408e539c0590c8fbdbed713fe5b30e//plugins/org.eclipse.sirius.tests.junit/src/org/eclipse/sirius/tests/unit/diagram/modeler/uml/LayerTests.java#L61