[test] Fix CompletionProposalInVSMTest tests based on property index
Since commit [1] and the change of Tycho version, the properties fields
order is no longer respected during the tests launched with Tycho. Thus,
tests depending on the order of the fields, like in
`propertiesBot.bot().text(2);` are no longer reliable.

This is the case for tests of CompletionProposalInVSMTest.

This commit used the tooltip to retrieve fields instead of their index.
The `testWithLabel()` method can not be used in the Properties view, as
the structure is not the expected one. Some tooltips, used in these
tests, are string constants, because at runtime they are a concatenation
of several other constants.

[1] https://github.com/eclipse-sirius/sirius-desktop/commit/08fb66d1d934bb7fa98329c42e21087bcc84306c

Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>