[fix] Replace wrong id for afterSection concerning label color
Replace afterSection="style.section.labelStyleDescription.LabelColor" by
afterSection="style.section.basicLabelStyleDescription.LabelColor"

The problem has been discovered since the commit [1] and the change of
Tycho version. When launching tests with this version of Tycho some
tests of TreeItemStyleDescriptionTest fail. The order of plugin.xml
loading seems to be change and it reveal the problem that exist since a
long time.
Observed problem: the property "Label Color" is displayed after the
property "Background Color".

[1] https://github.com/eclipse-sirius/sirius-desktop/commit/08fb66d1d934bb7fa98329c42e21087bcc84306c

Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>