[564] Fix the action Auto Size in contextual menu
Since the commit [1] about a memory leak fix, the Auto Size action is
the same instance for the diagram contextual menu, the "Diagram" main
menu and the main toolbar. Another instance is created only for the
diagram toolbar. Before that, there were many instances of the Auto Size
action.
A long time ago (2021), in commit [2], a clean-up was done to remove
actions contributed by GMF and to dispose them. Indeed, even if these
instances of actions were no longer visible, these multiple instances
were listening to all events and could have side effects.

The "dispose" of commit [2] and the fix of commit [1] are in conflict,
because there is now only one instance of the action that is disposed
and the recreated (but this is "not compliant" with the listening
mechanism and the recreated action is not correctly listening events).

[1] bcd26f29a137bbde1f3f4c5ee16fc00b5979c8bb
[2] 6bff02680031fa53d79e1f425327fa11db4efeb0

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/564
Signed-off-by: Séraphin Costa <seraphin.costa@obeosoft.com>