[test] Use a hard-coded font name for LabelFontModificationsTest
A previous attempt [1] was done to be compliant with all OS, by listing
the available fonts of the OS. Unfortunately, this doesn't work. It
appears that the fonts listed with
"GraphicsEnvironment.getLocalGraphicsEnvironment().getAvailableFontFamilyNames();"
correspond to the installed fonts for all users and the combo probably
displays the fonts installed for the user running the test.
This commit uses an hard-coded font name that is available everywhere:
"DejaVu Sans Mono"; at least, everywhere we tested it (locally on
Windows, Linux and Mac and on CI servers).

[1] https://github.com/eclipse-sirius/sirius-desktop/commit/e9a69fa28ede070e98ca133282ceedf4819dedc3

Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>