[609] Prevent Move request to reset size
Before this commit, nodes had their size changed when only a move was
requested.

An issue was detected by following case:
- Requires : Preference for 'Snap to grid for new diagram' is true
- Create a new diagram with synchronous mapping containing resizable
ports.

What is expected:
- the GMF elements are created.
- the ArrangeAll action is performed.

The previous and unexpected implementation:
- While GMF elements are created, AbstractCanonicalSynchronizer is call
to give initial positions and sizes.
- When ArrangeAll is performed, some new coordinates are computed and
sent to XYLayoutEditPolicy.
- AirXYLayoutEditPolicy implementation will receive ChangeBoundsRequest
to move the element but will create a SetBoundsCommand that change
position and size (set to -1, -1).

This modification ensures that when a ChangeBoundsRequest is emitted
only about moving element, the SetBoundsCommand only changed the
position and do not change the size.

Bug: #609
Signed-off-by: Nicolas PERANSIN nicolas.peransin@obeo.fr