[590] Prevent message from moving before the lifeline start
A condition allowed message move beyond the source lifeline range in
order to be able to move down a message destruction and resize both
lifelines. However this allowed message with a gate as target to be
moved before the beginning of the lifeline without a forbidden cursor. A
condition has
been added so the move only allow to move after the end of the lifeline.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/590
Change-Id: Ia1c13efa24ad5b080bc6183b31d360cb6eb07ae4
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>