[590] Fix connected Gate move that resized container
GateSelectionEditPolicy.getMoveCommand(ChangeBoundsRequest) was
cancelling the changeboundrequest to prevent a gate connected to a
message to be moved. However, it was not returning an
UnexecutableCommand and there was a bendpoint refresh that caused a
global refresh that in turn caused a resize of the parent.
showChangeBoundsFeedback has also been updated so trying to move a
connected gate won't display a "candidate position" shadow anymore.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/590
Change-Id: I9ab4f083fb5df7a609b29d99feacfe7764834cc7
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>