[590] No DoNothingCommand when createtool has no effect
Applying the combined fragment creation tool on an interaction use
(click center) has no effect. However a DoNothingCommand is executed.
This is added to the command stack and it result with having an undo
with also no effect.
This is a regression due to adding a compartment edit part for
InteractionUse that installs SequenceNodeCreationPolicy.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/587
Change-Id: Ida61c5285ae8f4aee09841b248b94397aabd0fa6
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>