[609] Update AbstractNodeSizeTest tests to check both snapToGrid states
Before this commit, there are regressions on CI server (not reproduced
locally). This is probably because of a default "SnapTo" configuration
that is different locally and on the CI.
All the tests of `NodeSizeOnDiagramCreationTest` failed with error `The
figure node Square_new EClass 1 do not have the expect height
expected:<200> but was:<204>`.

This commit:
* fixes a default snapToGrid behavior (no snap for NodeSizeTest and
NodeSizeOnDiagramCreationTest).
* adds similar classes but with the snap enabled
(NodeSizeWithSnapToGridTest and
NodeSizeOnDiagramCreationWithSnapToGridTest).
* changes the default size for container in vp973.odesign to be able to
see humanly the sizes and the potential problems. The vp973.aird has
been refreshed and elements moved to clearly see the result.
* ignores "Workspace image" style. This case is not handled with the
snapToGrid for border nodes because the ratio of the image has priority.
* replaces deprecated Sirius Option by Java Optional.

Bug: https://github.com/eclipse-sirius/sirius-desktop/issues/609
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>